% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crochet.R
\name{crochet}
\alias{crochet}
\title{crochet (first line)}
\usage{
crochet(
  rnw,
  df,
  folder,
  fnv,
  exp1 = NULL,
  exp2 = NULL,
  subset,
  III = NULL,
  ...
)
}
\arguments{
\item{rnw}{the rnw file}

\item{df}{a data frame}

\item{folder}{for the results (i.e feedback files)}

\item{fnv}{File name variable (a variable in data frame, df, to use as names of the resulting pdf files)}

\item{exp1}{expression to be evaluated within the data frame}

\item{exp2}{expression to be evaluated within the function's environment}

\item{subset}{subset}

\item{...}{passed to knit2pdf}
}
\value{
nothing useful, used for side fx
}
\description{
crochet (details here)
one
two
some more
}
