% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crochet_txt.R
\name{crochet_txt}
\alias{crochet_txt}
\title{Crochet-txt}
\usage{
crochet_txt(
  txt,
  df,
  exp1 = NULL,
  exp2 = NULL,
  subset,
  name = NULL,
  attrs = NULL,
  PS = "**",
  na.destroy = FALSE,
  browse = FALSE
)
}
\arguments{
\item{txt}{text}

\item{df}{data frame}

\item{exp1}{expression 1}

\item{exp2}{expression 2}

\item{subset}{subset}

\item{name}{name}

\item{attrs}{NULL}

\item{PS}{**}

\item{na.destroy}{Print empty string instead of NA}

\item{browse}{FALSE}
}
\description{
Crochet-txt
}
