% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboxplot.R
\name{mboxplot}
\alias{mboxplot}
\title{Mboxplot}
\usage{
mboxplot(
  x,
  Q = TRUE,
  TXT = "",
  BOX = "=",
  AT = NULL,
  jit = FALSE,
  XLIM = NULL,
  MAR = c(2, 8, 0, 0),
  MTEXT_CEX = 1,
  browse = FALSE
)
}
\arguments{
\item{x}{- a variable}

\item{Q}{(boolean) show quantiles by colors}

\item{TXT}{- text at the left side}

\item{BOX}{- box type "O" or "="}

\item{AT}{- passed to axis(1)}

\item{jit}{- jit points? (T / F)}

\item{XLIM}{- limits for x axis}

\item{MAR}{- margins (left one is large for text)}

\item{MTEXT_CEX}{- cex for label}

\item{browse}{}
}
\value{
a data frame
}
\description{
Mboxplot
}
