% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email - Copy.R, R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send e-mail using rJython}
\usage{
send_email(
  to,
  from,
  subject,
  message,
  attachment = NULL,
  username,
  password,
  server = "smtp.gmail.com:587",
  confirmBeforeSend = TRUE,
  wait = 0
)

send_email(
  to,
  from,
  subject,
  message,
  attachment = NULL,
  username,
  password,
  server = "smtp.gmail.com:587",
  confirmBeforeSend = TRUE,
  wait = 0
)
}
\arguments{
\item{to}{recipients name}

\item{from}{sender's name}

\item{subject}{subject}

\item{message}{message}

\item{attachment}{a file name}

\item{username}{user name to log in to the SMTP server}

\item{password}{password to log in to the SMTP server}

\item{server}{server (defaults to gmail SMTP server)}

\item{confirmBeforeSend}{ask before sending? defaults to true}

\item{wait}{Seconds to wait before sending next email}
}
\description{
Send e-mail using rJython

Send e-mail using rJython
}
