% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendit-package.R
\docType{package}
\name{sendit-package}
\alias{sendit-package}
\title{sendit-package}
\description{
\tabular{ll}{
Package: \tab sendit\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-03-29\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
Sendit
}
\examples{
NULL


}
\references{
\url{http://en.wikipedia.org/wiki/crochet}
}
\seealso{
\code{\link{crochet}}
}
\author{
Kenn Konstabel \email{kenn.konstabel AT psych.ee}
}
\keyword{package}
